+++ lib/dartlang_extensions.dart
@@ -1,0 +1,4 @@
+export 'src/extensions/StringExtensions.dart';
+export 'src/extensions/ListExtensions.dart';
+export 'src/extensions/NumberExtensions.dart';
+export 'src/extensions/BooleanExtensions.dart';