+++ lib/dartlang_utils.dart
@@ -1,3 +1,3 @@
-library plain_utils;
-
-export 'src/StringUtils.dart';
+export 'src/utils/StringUtils.dart';
+export 'src/utils/MathUtils.dart';
+export 'src/utils/BooleanUtils.dart';