+++ CHANGELOG.md
@@ -8,9 +8,32 @@
 ##### Improvements: 
 * Improved readme
 
+## [0.2.0-nullsafety.1] - 26.11.2020
+
+#### Fixes: 
+* Fix formatting
+
+## [0.2.0-nullsafety.0] - 26.11.2020
+
+#### New Features: 
+* Add Levenshtein distance to string utils
+* Add MathUtils
+* Add extension classes for bool, list and string.
+
+## [0.1.0-nullsafety.0] - 22.11.2020
+
+#### New Features: 
+* Support null safety
+
 ## [0.0.2] - 22.11.2020
 
-##### New Features: 
+#### New Features: 
+* Add hamming distance calculation to StringUtils
+* Add example to the package
+
+## [0.0.2] - 22.11.2020
+
+#### New Features: 
 * Add hamming distance calculation to StringUtils
 * Add example to the package